Skills/Py/TelegramBotAiogram4/requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
psycopg2==2.9.9
asyncpg==0.29.0